add static features
