+++ examples/experimental/weather/training_recipe/conf/afno.yaml
@@ -44,4 +44,7 @@ training:
 validation:
   batch_size: 1
 
+use_cos_zenith: False
+use_latlon: False
+
