+++ examples/experimental/weather/training_recipe/conf/config.yaml
@@ -27,21 +27,18 @@ training:
   num_steps: 2
   num_samples_per_year: 1456
   start_year: 1999 # TODO (mnabian): change
-
+ 
 validation:
   data_dir: "/code/dataset/validation/"
-  num_steps: 8
+  num_steps: 3
   num_samples_per_year: 4
   start_year: 1998 # TODO (mnabian): change
 
 #stats_dir: "/data/stats/"
-num_workers: 0  # TODO (mnabian) change to 8
 #stats_dir: "/data/stats/"
 lsm_filename: "./static/land_sea_mask.nc"
 geopotential_filename: "./static/geopotential.nc"
-#num_workers: 8
 
 dt: 6
-use_cos_zenith: False
-use_latlon: False
 stride: 1
+num_workers: 0 # TODO (mnabian) change to 8
